+++ src/org/sosy_lab/cpachecker/cpa/smg/graphs/CLangSMG.java
@@ -528,7 +528,7 @@ public class CLangSMG extends SMG {
     SMGEdgeHasValueFilter filter = SMGEdgeHasValueFilter.objectFilter(objectAtLocation);
 
     if (pLocation.isReference()) {
-      filter.filterAtOffset((int) pLocation.getOffset());
+      filter.filterAtOffset(pLocation.getOffset());
     }
 
     // Remember, edges may overlap with different types
@@ -581,13 +581,13 @@ public class CLangSMG extends SMG {
     }
 
     SMGObject object = initialRegion.get();
-    List<Integer> offsets = pLocation.getPathOffset();
+    List<Long> offsets = pLocation.getPathOffset();
     SMGEdgeHasValue hve;
-    Iterator<Integer> it = offsets.iterator();
+    Iterator<Long> it = offsets.iterator();
 
     while (it.hasNext()) {
 
-      int offset = it.next();
+      long offset = it.next();
       Set<SMGEdgeHasValue> hves =
           getHVEdges(SMGEdgeHasValueFilter.objectFilter(object).filterAtOffset(offset));
 
@@ -762,14 +762,14 @@ public class CLangSMG extends SMG {
       Integer pLocationOnStack, String pVariableName) {
 
     Set<SMGEdgeHasValue> objectHves = getHVEdges(SMGEdgeHasValueFilter.objectFilter(pSmgObject));
-    List<Integer> offsets = new ArrayList<>();
-    Map<Integer, SMGObject> offsetToRegion = new HashMap<>();
-    Map<Integer, SMGMemoryPath> offsetToParent = new HashMap<>();
+    List<Long> offsets = new ArrayList<>();
+    Map<Long, SMGObject> offsetToRegion = new HashMap<>();
+    Map<Long, SMGMemoryPath> offsetToParent = new HashMap<>();
 
 
     for (SMGEdgeHasValue objectHve : objectHves) {
       Integer value = objectHve.getValue();
-      Integer offset = objectHve.getOffset();
+      long offset = objectHve.getOffset();
 
       SMGMemoryPath path =
           getSMGMemoryPath(pVariableName, offset, pPos, pFunctionName, pLocationOnStack, pParent);
@@ -789,7 +789,7 @@ public class CLangSMG extends SMG {
 
     Collections.sort(offsets);
 
-    for (Integer offset : offsets) {
+    for (long offset : offsets) {
 
       SMGObject smgObject = offsetToRegion.get(offset);
       SMGMemoryPath currentPath = offsetToParent.get(offset);
@@ -798,7 +798,7 @@ public class CLangSMG extends SMG {
     }
   }
 
-  private SMGMemoryPath getSMGMemoryPath(String pVariableName, Integer pOffset,
+  private SMGMemoryPath getSMGMemoryPath(String pVariableName, long pOffset,
       SMGObjectPosition pPos, String pFunctionName, Integer pLocationOnStack,
       SMGMemoryPath pParent) {
 
@@ -903,9 +903,9 @@ public class CLangSMG extends SMG {
       Integer pLocationOnStack, String pVariableName) {
 
     Set<SMGEdgeHasValue> objectHves = getHVEdges(SMGEdgeHasValueFilter.objectFilter(pSmgObject));
-    List<Integer> offsets = new ArrayList<>();
-    Map<Integer, SMGObject> offsetToRegion = new HashMap<>();
-    Map<Integer, SMGMemoryPath> offsetToParent = new HashMap<>();
+    List<Long> offsets = new ArrayList<>();
+    Map<Long, SMGObject> offsetToRegion = new HashMap<>();
+    Map<Long, SMGMemoryPath> offsetToParent = new HashMap<>();
 
 
     for (SMGEdgeHasValue objectHve : objectHves) {
@@ -916,7 +916,7 @@ public class CLangSMG extends SMG {
       }
 
       SMGObject rObject = getObjectPointedBy(value);
-      Integer offset = objectHve.getOffset();
+      long offset = objectHve.getOffset();
 
       if (!isHeapObject(rObject) || pReached.contains(rObject)) {
         continue;
@@ -935,7 +935,7 @@ public class CLangSMG extends SMG {
 
     Collections.sort(offsets);
 
-    for (Integer offset : offsets) {
+    for (long offset : offsets) {
 
       SMGObject smgObject = offsetToRegion.get(offset);
       SMGMemoryPath currentPath = offsetToParent.get(offset);
